add checks to build script

+++ build.ps1
@@ -1,5 +1,8 @@
 $outputPath= [IO.Path]::Combine($pwd,"artifacts")
-Remove-Item $outputPath -recurse
+If (Test-Path $outputPath)
+{
+  Remove-Item $outputPath -recurse
+}
 New-Item -Itemtype directory $outputPath -Force
 $msbuild="$($env:systemroot)\Microsoft.NET\Framework\v4.0.30319\MSBuild.exe"
 $nuget=".\.nuget\nuget.exe"
@@ -9,5 +12,11 @@ Write-Host -ForegroundColor Yellow "Restoring Nuget Packages..."
 &$nuget restore
 Write-Host -ForegroundColor Yellow "Done."
 
+$verbosity = "minimal"
 
-&$msbuild $sln /m /nologo "/p:configuration=release;OutDir=$($outputPath);VisualStudioVersion=14.0" /t:rebuild
+if($env:MSBUILD_VERBOSITY)
+{
+  $verbosity = $env:MSBUILD_VERBOSITY
+}
+
+&$msbuild $sln /m /nologo "/p:configuration=release;OutDir=$($outputPath);VisualStudioVersion=14.0" /t:rebuild /verbosity:$($verbosity)
